third_party/blink/web_tests/external/wpt/css/css-anchor-position/anchor-name-in-shadow-expected.txt

This is a testharness.js-based test.
FAIL anchor-name should not leak out of a shadow tree assert_equals: #anchored is positioned against #anchor expected 100 but got 0
FAIL anchor() in shadow tree should not match host anchor-name assert_equals: #anchored is positioned using fallback expected 37 but got 8
Harness: the test ran to completion.
